Don't destroy &statusline; autoclose quickfix
There was a bug where after an :AsyncRun command finishes I was trying
to force Vim to redraw the statusline by doing let &statusline = &statusline.
Unfortunately this effectively copies the window-local option into the
global option, and the window-local option happens to be one for the
quickfix statusline!

Also, I want the :AsyncRun callback to close the quickfix window it had
opened on a previous run, if the previous run was a failure, but the
current run is a success.  To keep thing simple I'll just close the
quickfix every time the run is a success.