Re-enable xterm title setting under Wayland
So gnome-terminal is a native Wayland app, so vim can't use the
X11-based hack for reading the window title.

I want the xterm title to show me the current file I'm editing.  I don't
mind not restoring the title because my PS1 will set it as soon as vim
exits.

(Sometimes vim is not launched from a shell, but that's mostly vim
launched from mutt, which has its own option to set xterm titles.)