:TermRestart for when I use vim as a tmux replacement
So I sometimes do things like vim +'term make run-server' and when the
command fails because autoreloading + Python syntax errors + saving
reflex when I'm in the middle of a refactoring causes it to fail, I want
to re-start it in the same split.

Annoyance: now :Te<tab> doesn't expand to :TestForTheOtherWindow any
more.