Fix statusline highlighting when quickfix (dis)appears
Use a big hammer to do everything possible, but I think I was just
missing the _update() call.

I think it's related to autocommands-not-triggered-from-other-autocommands
rule, where I open/close quickfix from an autocommand and so statusline
doesn't get a chance to update the active/inactive statuslines for the
right windows.