ISO-8601 datetimes are not filenames, Vim!
I like to run Python unit tests and look at the results in the quickfix
window.  One thing that annoys me is that when the output contains
assertion failures involving ISO-8601 timestamps like this one:

    E       AssertionError: assert 'GitLab proje...tion:   42s\n' == 'GitLab proje...ob created:\n'
    E           GitLab project: owner/project
    E         - Job created: tbd
    E         + Job created:    2020-09-16T06:16:49.180Z
    E         + Job started:    2020-09-16T06:16:51.066Z
    E         + Job finished:   not yet
    E         + Job duration:   42s

    tests.py:346: AssertionError

Vim thinks there's an error on line 16 of a file named
"E         + Job created:    2020-09-16T06"

There's also the same kind of issue with the actual source code of the
function, only without the leading E.

Gary Johnson suggested adding an &efm entry early on that would match
these lines and ignore them.  Here's the regexp I picked:

  ^.  .*$

(AFAIU the ^ and $ are implicit, and other metacharacters need some
translation for &efm syntax, namely . -> %. and * -> %#.)