Snippet to remind me about pytest.raises syntax
The bits I keep forgetting are name of the `match` argument and the
name of the `value` attribute on the context manager that contains the
caught exception object.

(Also, I'm uncertain about the best name for the context manager itself,
and now I see pytest's documentation examples use `excinfo` for it.)

(Is it the context manager or the exception information object returned
by the context manager?  Does it matter?)