Improve magical shiftwidth for strings
I want 2-space indents for YAML markup inside docstrings/long
tripple-quoted literals, but I don't want 2-space indents _ever_ for the
opening/closing line that has the tripple quote marks.

The condition currently is a first approximation that fixes the cases
I was annoyed about, namely pasting a bunch of text like

    lines = [
    "aaaaa",
    "bbbbb",
    "cccc",
    ]

and trying to indent them all in visual block mode.  The cursor being on
column 1 of the "aaaa" line counts as it being inside a string, which
used to mistakenly set sw=2, after which it was outside the string and
reset sw=4, meaning I'd get indentations of 2, 6, 10, 14 etc instead of
4, 8, 12, 16.